Catch missing test symbols
`str_match` is a bit delicate in how it matches function names, and if
a slightly misformed line results in a non-valid symbol name, the test
runner segfaults.

This catches the case of the function name not actually being a valid
symbol in the shared object file, and "fails" the test with a "Skipped"
message about not begin able to find the symbol.